refactor(memory): move activation types to activation.go
The volatility fields pushed store.go past its recorded budget; the
activation concern gets its own file (the freshness.go/render.go
precedent) instead of a baseline bump.
